* feat/PlayerHP:
  feat: 調整減傷機制計算公式、新增開場時選擇三選一武器的設計
  feat: 增加玩家防禦力設計、調整槍械傷害系統
  feat: 調整升級選項的生命回復機制
  style: 移除近戰武器相關程式及預製物
  feat: 調整玩家血量顯示模式

+++ Assets/Scripts/Game/Interface/IGameUIController.cs
@@ -10,7 +10,7 @@ namespace Scripts.Game
 
         public void GetExp(ExpNumber exp);
 
-        public void HealPlayer(int healPoint);
+        public void HealPlayer(float healPoint);
 
         public void AddPlayerHealthPointMax(int healPoint);
     }
